requirements.txt
absl-py==2.0.0
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2023.4.0
gast==0.5.4
gdown==4.7.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.4
Jinja2==3.1.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==9.3.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
six==1.16.0
some-package==0.1
soupsieve==2.5
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==3.0.1
wrapt==1.14.1
